File API
Super happy about this! While working on the cross-platform (in regards to the different Minecraft versions/platforms) Minecraft world APIs, I decided to go back and rework the code for this projectt before moving this existing code over to there. I looked into the status of the File API constructor in NodeJS, as I remembered it being one of the most recent things that would be a great upgrade here, and looks like it just landed about a month ago! Super awesome. That takes one more thing out, and now it's a little bit less of a headache to make work everywhere. That's awesome.

https://github.com/nodejs/node/issues/39015
https://github.com/nodejs/node/pull/47153
https://github.com/nodejs/node/commit/7bc0e6a4e79684a98cb64c61452af00e20f590e7