Generator Iterators (T1000)
Alright, took the last few days to get this fully working, and I can fully say that this is the best version of the codebase so far. Everything is built using Generators, rather than arrays of the individual items themselves. I really like this because it doesn't require you to get all of the items in order to get a specific item. At least, if you find it before you get to the end of the Generator. I think it better represents what the archive traversal does though, so it's more to help with conceptualizing that instead of in terms of a performance gain.

While building the thing from the ground up yet again, I started to see some more natural places where breaking everything into separate modules started to make more sense than I had seen before.

The best discovery tonight from the original codebase that I seem to have been unable to locate for pretty much the whole time I've been working on it, is the reading offset itself for calculating `"File Information".Filesize` and `Offset`. Every time I have tried to rewrite the project to make it better, I think I kept forgetting to include reading past the `Filename` bytes' length `0x80` when trying to get those values for the definition entry. This in turn would read those values out of the name bytes itself, making up incorrect size and byte offsets for the resulting files. This was hard to discern because when doing the live testing to check for the output results, I would think it would parse correctly when the name worked, but I wasn't looking closely enough at the file size details of the output, and whether they had the correct data in them (and, opening correctly with NBT editors). I'm glad I have tried adding constant declarations for different sizes for the different parts of the format, it really does help explain what's going on and why you are adding some values when you use them. I'm going to try this with my other projects too, I really like the explicit nature that it provides when looking things over again later. You have a better idea of *why* you did what you did, not only what you did. Having the *why* part is possibly equally, or more important than the what.

https://github.com/PhoenixARC/Minecraft-LCE-Docs/blob/main/Documentation.md

Another thing, I have found that finding out where you want to break up your project into modules helps really well if you do it by separation of concerns, rather than just to 'organize' things. I have found that using it as a tool to define what things can work with other things is much more helpful, because it also makes it explicit as to what each thing needs to do it's job. It helps show what it's job is too.