Compact Definition Parsing
After doing some experimenting over in MCRegionJS, I realized how I could implement the same chunk parsing logic using a single Generator to read the data, rather than having one to get the chunk, and one to read the chunk itself. With one Generator, I can just iterate over the block of bytes for the file itself as a whole, and read from specific offsets instead. I think this could possibly be more performant because it doesn't allocate new TypedArray views for each chunk itself, you just read the data from the chunks at those given offsets once you get to that part of the iteration step.

It's a little further away from the current implementations for what I have built so far, but this could be a nice way to get specific chunks/regions, things like that, since you can't necessarily get specific items when accessing them through an iterator.

https://stackoverflow.com/questions/30410359/how-to-get-the-nth-value-of-a-javascript-generator

I want to allow the API (for working with all of these different world data types) to provide the user with a way to get a specific chunk (or region) if they need to edit just a single one.

Looks like the output NBT files are correctly opening with the VSCode NBT editor extension, and webNBT, but NBTify doesn't appear to be handling things correctly for some reason. It's encountering unexpected bytes in a few spots a far way into the file. Wanted to get this added before debugging that next too.