Decrypted PS4 Save Support!
Wow! Debugged getting PS4 save support working! It uses a little different of a format to that of a PS3 save.

So it takes the whole `GAMEDATA` file, deflate-compresses that, then prefixes that with a header of the decompressed payload's length. The original `GAMEDATA` header bytes are then inside the now compressed part of the file.

Non-Region files are now found with in the Dimension folders, looks like LCE gained the separated support of storing entities in their own files, outside of the Regions themselves. My Nether Dimension folder name handling now accounts for files non-prefixed with the `r-` naming scheme, allowing them to properly go into the directory again, rather than be written to the root of the world folder with the `DIM-1` prefix (existing issue with PS3 saves too).

The `GAMEDATA` archive header *and* file definitions now use little-endian for the file count length, byte offset, and byte length values, and the file names now use UTF-16 little endian rather than big-endian. So looks like across the board they moved to little-endian, previously from being big-endian across the board.

Thanks to EternalModz for providing me with this decrypted world to test on!

So right now there isn't a user-accessible way to decrypt worlds without jailbreaking your PS4, but at least if you do go through with doing that, you will be able to edit your world/convert it to other versions, the tools will be able to support those game versions. It's just a matter of decrypting the world to be able to edit it, whether or not the world version/PS4 platform is supported is now out of the question :D (kind of, I guess)

Thanks to HexEd.it for debugging the contents of the file! Noticing the zlib header made things start to feel less hopeless than they were (I was ready to give up and not worry about PS4 support)

https://hexed.it/
https://stackoverflow.com/questions/9050260/what-does-a-zlib-header-look-like (and thanks to my nerdy brain for being able to notice the zlib header just by memory haha, can't believe that's just on the top of my head now haha; had to Google it again to make sure I wasn't just hallucinating it)
https://github.com/UtterEvergreen1/Minecraft-LCE-Docs/blob/main/Documentation.md (Used to debug whether the `GAMEDATA` format itself changed; turns out it was just the endian change that made it not parse, I initially thought the `GAMEDATA` format had changed, causing it not to parse)