Full PS4 World Data
Not totally sure how the file structure of these files look, I received them in two zip files under these two names with the `0000000-*` prefix. Including them like this here to start with, in case this is intended to have them organized this way. Looks like the only difference between the two are the extra `GAMEDATA` files, the folder structure is the exact same, and the `sce_sys` folders have the same exact content too.

The current idea for as to what these additional numerically-suffixed `GAMEDATA` files are is that they might be the Region files that are missing from the regular `GAMEDATA` file, which usually includes them as of the PS3 `GAMEDATA` format.

*Edit: Decided to remove the duplicate `sce_sys` folder from the `00000006` folder since Git doesn't really need to track them. They are the same exact files as those in the `00000005` folder. The only difference between the new `00000006` and the original content from the `00000005` folder are the `GAMEDATA` entries. `00000005` was the original one I added to the repo at the start of this delve into experimentation.