PS4 Data Organization
Just restructured the demo folders a bit since I don't think the folder structure was quite intended with how they were zipped. I'm not completely sure on that though, so that's why I wanted to track that along with the repo.

Working on my Lordsmith release tonight, doing a nice listen-through. Zeke :)