'Region Files' Demo
Trying to deduce what the heck these additional `GAMEDATA_*` files are! I'm pretty sure they are the missing `.mcr` files that are usually present in the standalone `GAMEDATA` file, but I'm not sure how you are supposed to distinguish them apart. Do they hold their file name anywhere?

My initial guess is that the start of the file stores the `r.*.*.mcr` numbers, but I'm not totally sure. I'm just guessing here, trying things out. The current version is just because the values it return appear to kind of look like what we might want to find.