File-System-Access (Node & browser?)
I know how to use the File System Access API, as well as Node's FS module, the hard thing is implementing both so this project can work in the browser directly, or on the terminal/locally. Then I thought it might make sense to use the FSA API structures across the board to manage things, no matter where the program is running, but I'm not sure this is fully the best way to go. I'm going to look into other options.

My first idea that I started working on, but started to feel messy, was working with `File` objects directly. That seems hard to manage though, since you have to build and tear them apart yourself. It's the next best thing to working right in the browser though, so I'm trying to figure out which way to go here. I may end up making my own polyfill/library sort of thing to manage this. It could be way easier, I just haven't found what I'm looking for yet.

https://developer.mozilla.org/en-US/docs/Web/API/File_System_API
https://developer.mozilla.org/en-US/docs/Web/API/StorageManager/getDirectory
https://developer.mozilla.org/en-US/docs/Web/API/Navigator/storage
https://developer.chrome.com/articles/file-system-access/#creating-or-accessing-files-and-folders-in-a-directory
https://developer.chrome.com/articles/file-system-access/#accessing-the-origin-private-file-system (Inspiration for bringing this implementation to both Node and the browser, by making the APIs the same to use under the hood)
https://www.npmjs.com/package/file-system-access
https://github.com/use-strict/file-system-access