LZX Implementations (Discovery)
Talked with Evergreen some more about how to get an LZX implementation set up, and turns out it's (only?) available through the archived Xbox SDK, which he helpfully provided me with the links to find.

https://www.retroreversing.com/xbox-sdk-xdk
https://archive.org/details/xbox-sdks

Going to take a break from reading the new format findings, 360 and PS4, as I want to start working on the intermediate format's implementation, maybe starting with Bedrock and Java now. I think that will make it easier to decide on how I should structure the APIs for reading worlds and such, as there will be more of an idea as to what the return types and such should look/work like. By that I mean dealing with NBT structs and Block and Entity data kinds of things.