Platform Test Runner
Going to go the test-based implementation route, which I learned with NBTify. You can more easily and quickly see when your code breaks, as well as works, for each individual case on it's own, rather than having to manually make sure to check each one on it's own.

I think I also wanted to set this up because I'm going to be rewriting the implementation for this, and I want to ensure it works the same as it did before, or rather one feature doesn't override another, because sometimes one can accidentally cancel out another one.