GAMEDATA Rehash
Worked on refactoring the `GAMEDATA` parser implementation for it more cleanly represents the parsing process. I thought originally that it might make sense to allow the API to only load certain files that you need, rather than having to get all of the files, but I'm not sure that's completely necessary. Maybe I can make that later as a simple step on top of the API. I don't think it has to be implemented from the ground up though. If I find the need of having that later on, then it could make sense to add it.

Moved the PS4 compression handling over to NBTify's implementation over the Compression Streams API, so it's not Node-specific anymore.

Realized that the files names are probably null-terminated, which is what most languages use to signify the end of strings. Since the empty bytes after the string contents are just `\0` (null), it makes sense that it would be the case to parse them in that manner. So it's a length limit of 128, but you know it's the end of the string because of the first `\0` character that follows.

I also renamed the `360` platform to `xbox-360` because the possibly inclusion of `one` wouldn't quite make sense. Neither would `u` haha, lol.