GAMEDATA Writing Test Runner
It's not implemented yet. I want to add it to the test runner though, to validate when it does seem to be writing symmetrically. I set this up earlier today, but I got lost trying to figure out more of the `GAMEDATA` format itself, so I didn't get to implement it yet. The issue is figuring out the size of the file from the files you provide to the writer. That's less the issue, but moreso that the start of the file appears to have some offsets/size values of some kind there, but we don't know what they are yet.

Reminder (to me too):
There docs are the ones to go to! They seem to be the most complete, or rather formatted/organized ones, of the different versions that are out there now. I'm not sure they are the most up to date though, that's the only issue to possibly be wary of.
https://modifiedcommand.github.io/Legacy-Docs/