Decompressed Console Debugging!
Wow, I was very surprised with this! Both the decompressed versions of the 360 and Wii-U file save formats are completely (from the looks of it) compatible with my parser. This is great! So I just have to figure out the compression implementation parts I think. I thought there was more differences between each of the file types, but it might just be general compression and headers are the only things. Both of these new files work with the test! Going to enable them as part of the test runner next. Just wanted to diff things properly.

https://modifiedcommand.github.io/Legacy-Docs/Saves/#minecraft-legacy-save-files

https://www.youtube.com/playlist?list=OLAK5uy_k1kzlhUdKcYYAlWvMQlYIHAEz_hFrJNQM

I'm starting to really like and click with Dust Speck. I think I needed a little break from Mike and his shenanigans (in the most kind way speaking 😂 I think it was getting too stuck in me, sometimes a refresher really helps with liking more of something). I already really liked this album for different parts, but I think I have more of the headroom for it to click again.