Compressed Save File Tests
`savegame.dat` is a new file, the one that was there before is now `-first`. I'm not certain that the new `savegame.dat` file is the compressed equivalent or not. I will be able to find that out eventually though.

Oh yeah, forgot to mention, I think this earlier for working on committing things in the proper order. Essentially I want there to be a `git stash pop --force`, but it's not a thing.

https://stackoverflow.com/questions/16606203/force-git-stash-to-overwrite-added-files