PS Vita World Test
Curious about what compression the top-level of a PS Vita save file uses.

I tried zlib, gzip, raw deflate, and RLE, and none of them work.

The file itself doesn't appear to have any of the regular appearances for any of those either, but I thought I might as well try by way of looping through the whole file to see if it could find any compressed parts in those formats.

Not sure whether the docs for this are accurate, but it doesn't seem to be lining up unfortunately.
https://github.com/PhoenixARC/Minecraft-LCE-Docs/blob/main/Documentation.md#lce-buffers