Other Vita Tests
Not sure the original one we were testing on was a Vita save (360, instead? Not sure, but it didn't seem to follow the compression we thought), so I got some other examples from Jerrin and Alexis.

Started working on a prototype implementation for Vita RLE, it's not fully known-working yet. Just going to commit all of the work I have here.

https://github.com/zugebot/LegacyEditor/blob/master/LegacyEditor/utils/RLE/rle_vita.hpp
https://www.tutorialspoint.com/c_standard_library/c_function_memset.htm
https://en.wikipedia.org/wiki/List_of_file_signatures