Vita Archive Entry Debunk
Reverse engineered figuring out how to get name entries from the Vita GAMEDATA file, by way of how the PS3 version works. It's not dynamic yet, but now we know that the names at the end are partially in the same format.

As of now, the files' contents don't match accordingly, I think because it's not reading the correct sizes/offsets from the file, so it's putting junk data in each file.

This is the concept I thought of to get to this point:

I think I'm going to figure it out backwards
I'm going to see what the byte offset to the first read data is for a PS3 save, like where to jump to for the names
I'm going to see where that spot looks familiar in the Vita save we have, then manually get the offset with the hex viewer
Then I'm going to encode that number into a buffer and see what the bytes for that look like
Then I'm gonna search the Vita file to see where it stores that number