Vita RLE Header Debug
The first 8 bytes are the compression info of the file, then the rest is RLE encoded.

https://github.com/zugebot/LegacyEditor/blob/181654ea59dc7c96d8cdfef86df33c47c580c1a1/LegacyEditor/LCE/FileListing/fileListingReader.cpp#L373

Decoding these makes the result content the same shapes between PS3 and the other Vita demo save!

It's still not properly reading the file content or the the names, I think it's something with my script. Dex's project seems to be working with this discovery though, going to debug my setup.

https://github.com/DexrnZacAttack/DexrnZacAttack.github.io/commit/ada877aea0aeddc8f9085ccdb48583df9e6a7c67