Vita RLE Improvements
Adjusted the logic behind the Vita RLE implementation, just inlined a few different things.

I also brought up the performance of it a lot, because I backed it with a fixed-length Uint8Array now that we know where the decompressed length is stored in the file! I think it was like 100ms before, now it's only 10ms!